Fix for modified QUrl::toLocalFile behavior
See https://doc.qt.io/archives/qt-4.8/qurl.html#toLocalFile:

"If this is a relative URL, in Qt 4.x this function returns the path to maintain backward compatability. This will change from 5.0 onwards. Then the path is returned only for URLs where the scheme is "file", and for all other URLs an empty string is returned."