Update from Qt 4 to Qt 5 (#3035)
* Wholesale qt4 to qt5 switch in all configs/shell scripts/etc.

* Removed deprecated qtestlib and script from project files

* Ben's Qt5 provisioning changes

from be9efa9 minus the port changes

* More of Ben's Qt5 provisioning changes

* Undid some of the provisioning changes

The vagrant VM was not properly built, process ended early. 
With this revert it looks like it's compiling now. On to the Qt5 code changes...
Include folder appears ok so far without qt4 stuff in it.

* Replaced all toAscii() calls with toLatin1()

* hoot-core and hoot-core test code changes for Qt5

- replaced fromAscii() with fromLatin1()
- use of new QUrlQuery class where needed
- createUuid return type change fix
- added QNetworkCookie include wherever needed for templates
- qendian include path fix
- qInstallMessageHandler interface change fix
- QtConcurrent include name change
- change QTextCodec function codecForCStrings to codecForLocale

* Project and code changes for Qt5

Changed QApplication include to QCoreApplication
Add concurrent dependency to hoot-cmd

* Added concurrent lib to tgs project

Everything builds, links and runs now. But tests fail big time...

* Added qt5 generated makefiles and qmake.stash to gitignore

* Force qt hash seed value to 0 for hoot-cmd and hoot-test

* OsmXmlWriter close() bug fix

Fixed an issue where the OsmXmlWriter descructor would try to write to an already closed file.

* Start updating quick test outputs

* Added missing != operator overload to Tags class

* New png test image for NodeDensityPlotCmdTest

* New results for PertyRemoveTagVisitor test

* RemoveRef2Visitor test result updates for qt5

* Fix for modified QUrl::toLocalFile behavior

See https://doc.qt.io/archives/qt-4.8/qurl.html#toLocalFile:

"If this is a relative URL, in Qt 4.x this function returns the path to maintain backward compatability. This will change from 5.0 onwards. Then the path is returned only for URLs where the scheme is "file", and for all other URLs an empty string is returned."

* Updated geojson test files

These files could be verified with hoot diff and/or compare

* Test file updates

Most issues are tag order
Exponentials no longer have a leading 0

* More updated test files

Looks like QPainter has changed and the render results are a different but comparison results are actual higher than before, so I assume this is ok.

* Reversed wrong TransliterateNameVisitorTest with arabic text

Reversed a test that assumed an arabic text was latin. Not sure why. Qt5 interprets it correctly as non-latin. This concludes the fixes for the --quick tests.

* Slow test output file updates

Tag order change
Also, tag "error:circular" is no longer created. This also happens with qt4 version so must be a remnant of on older version of hoot

* Numerous test updates for slow hoot-test process

* alt_types= value bug fix

Fixed a bug where 'alt_types=' would be added to the value of an alt_types tag

* Updated fix for alt_tags handling, fixes qt5 exposed issue as well

* Restore vagrant ports to default values

* Glacial test result updates for qt5

* Fix for MultiaryReviewCommandTestE::hashTest

* Enforce hash seed 0 for js interface/plugin calls

* TranslationServer.js tests tag order fixes

* Fix for majority of js unit tests for Qt5

* Temporarily disabled two JS TDSv61 translation tests

Temporarily disabled two JS TDSv61 translation tests until the tag sorting issue impacting the translation results has been resolved.

* Copyright updates

* Enabled verbose build in Jenkinsfile

* Update build images, including an image of `rpmbuild-hoot-release` that includes Qt5.

* Added hash.seed.zero config option for testing

Removed hardcoded seed in cmd main
Set config option to true and seed to 0 in test main

* Added '-D hash.seed.zero=true' to shell script based tests

* Plugin/js test support for setting hash seed to 0

Also fix for a test shell script that was missed before

* Copyright update

* Revert "Merge remote-tracking branch 'origin/develop' into qt5_2"

This reverts commit 82fc6bcfd3239330a981361c397559a897f8a6df.

* Script test fixes

* Cleaned up QNetworkCookie include, moving it to HootNetworkCookieJar.h

* Copyright fix

* Centralized testing specific options in Testing.conf file

* Test file conf reference fix, missed earlier

* Revert "Revert "Merge remote-tracking branch 'origin/develop' into qt5_2""

This reverts commit 0d8757b3fa157e03ea3ac7aefddc9aed9a58b2c1.

* Revert "Enabled verbose build in Jenkinsfile"

This reverts commit 5dfc992c118f43022fe82536d734d5e3d065d525.