remove incoming adv opt fix that we no longer want (#3112)
Rubbersheeting with the Unifying alg was originally disabled b/c rubbersheeting was occurring, even when unselected, from the UI.  Now, some users would like to use it with the Unifying alg so removing the hardcode that disabled it.  If it still is being applied regardless, then we can open a UI bug issue to fix.