fix json big into parsing issue (#3165)
The regex used to replace large integers in JSON values to address Boost JSON parser bug was causing certain large integers embedded in JSON text values not to parse (see OsmJsonReaderTest::scrubBigIntsTest). A simple solution was to slightly relax the regex by removing a semicolon to match, as no test data has been found that needs that level of strictness. If such data exists, then a different solution would need to be found.