Fix bug where UFD buildings with ref overlap where being returned in conflate diff output (#3210)
Removed the pre-conflate code in Differential Conflation that drops all relations from input. This was causing UFD secondary buildings to be reduced to building parts only and, therefore, they couldn't match with anything in the ref layer. Diff conflate is based entirely off of match conflicts, so with no matches there was no real diff and overlapping UFD buildings were incorrectly added to the output. I'm not sure what the original purpose of dropping the relations was, but keeping them doesn't break any tests...so going with that as the fix.