Allow ElementVisitor to be used from JS scripts (#3279)
Prior to this change only `ConstElementVisitor` could be used from a JS script due to how the hoot-js library was initially designed. Now both it and `ElementVisitor` can be used in the scripts. Unfortunately, to make this change there was a cascading set of const correctness related changes needed to make it happen, culminating in a bad const cast in `FilteredVisitor`. The capability added far outweighs the single bad change, and likely some redesign work is in order to make that cast go away.