Finish implementation for changeset.allow.deleting.reference.features option (#3347)
The original implemenation was only partially completed for an unknown reason. This option is particularly helpful when generating changesets for Differential Conflation output when you don't want to change any reference features. Given that, technically, we could implement something similar to suppress modify statements but haven't seen a use case for that yet.