1058 - score-matches-diff command (#3520)
Adds a command, `score-matches-diff`, which compares the outputs of two invocations of the `score-matches` command against the same input data. 

The `score-matches` command provides basic information about the accuracy of a conflation job when compared to ground truth (manual matches). However, it is difficult to compare the output of two separate runs of `score-matches` against the same dataset. `score-matches-diff` allows for that and is useful when you want to see detail on how matches changed when tweaking conflate configuration options or logic.

* Added `--input-streamable` and `--output-streamable` as options to `info --formats`. These will display supported streamable input and output formats.
* Added `ElementIdToTagValueMapper`, which allows for collecting an association between element IDs and tag values for all particular elements that have a specified tag key.