1063 - Robust manual match error checking (#3523)
Prior to these changes the error checking for the manual matches used to score hoot conflation in the regression tests was pretty minimal. The absence of good error checking meant that if a manual match mistake was made it was fairly difficult to track down and often would go unnoticed, thus leading to inaccurate match scoring. The new validation runs automatically every time `score-matches` is run and will exit the command with an error status if any validation error is found. The first ten errors are output to the display.