reduce test parallel threads from 8 to 6; core tests have been going deep into swap and slowing down unnecessarily for some time and recently running out of memory periodically; some of the tests require extra memory now and that will have to be increased on the test vm to run in full parallel (#3810)
The core tests have been going deep into swap and slowing down unnecessarily for some time now and recently running out of memory periodically. Some of the tests require extra memory now and the overall memory will have to be increased on the test vm to run in full parallel. This solves the issue for now. We can go back and try to increase the value back to full parallel at some point in the future after looking more closely at the tests.