3713 - Tailor conflate ops to type of conflation being run (#3796)
Checks have been added in `ConflateCmd` to filter out operators in `conflate.pre.ops`, `conflate.post.ops`, and `map.cleaner.transforms` (if `MapCleaner` is specified in either of the aforementioned op lists) when they don't apply to the configured conflate matchers in the `match.creators` option before a conflation job is launched. For example, if only building conflation is being run all references to road specific cleaning conflate ops, such as divided highway splitting, etc. are automatically removed. This can result in increased runtime performance for conflation jobs where matchers are custom configured.

This was accomplished by adding an interface, `FilteredByCriteria`, and implementing it on all `MatchCreator` classes and select `OsmMapOperation` and `ElementVisitor` classes. The method `FilteredByCriteria::getCriteria` returns a list of all the criteria used by the matcher, op, or vis when selecting feature to operate on. The list is logically OR'd together to create the feature criterion. On `MatchCreator` implementation of the interface is required. Script matchers called by `ScriptMatchCreator` must supply the `exports.matchCandidateCriterion` property with the criteria listed in a semi-colon delimited list. On `OsmMapOperation` and `ElementVisitor` implementation is optional, and if not implemented, its assumed the op or vis is meant to run against all feature types.

As a failsafe (maybe temporary) the option, `conflate.remove.superfluous.ops`, enabled by default, allows for turning off the removal of the superfluous ops before conflation.

Also, a handful of ops that operated only on roads were renamed to reflect that fact.