3802 - Changeset replacement derivation not removing features correctly from reference with --full-replacement option (#3817)
Fixed a bug where if you define `--full-replacement` with `changeset-derive-replacement`, the reference map contains feature types absent in the secondary replacement dataset, those feature types weren't being deleted in the resulting changeset.

e.g. building and road in reference dataset, road only in secondary dataset; with `--full-replacement` the building should be deleted

The behavior remains the same as before when `--full-replacement` is not specified, as there are no corresponding features of the same type in secondary dataset as in the reference dataset to do overlapping only replacement.