3815 - Remove type=multilinestring only relation restriction on road conflation (#3822)
There had been a restriction added to not allow conflating roads in any relation unless that relation had type=multilinestring. A user came across a type=route relation and was unable to conflate it. I removed the restriction, made another unrelated small change, and the roads appear to conflate fairly well. Since I don't the history behind the restriction, I left a note in the code with it commented out explaining that we could re-enable it in the future if we found related relation road data with a type not equal to multilinestring that hoot doesn't conflate well.