3828 - Improve stats command progress reporting (#3836)
While trying to collect map stats for #3814, I noticed that the stats are numerous enough now that its difficult to follow progress when they're run on large datasets. Updated the logging in `CalculateStatsOp` to convey better progress information.

Also moved the `OperationStatusInfo` interface up to the parent classes: `ElementVisitor` and `OsmMapOperation`. This makes the code a little cleaner. Updated `OperationStatusInfo` and added it to a few classes used by the stats. Really, all child classes should be implementing but will save that for a future task and allow an empty default implementation due to the amount of work involved.