Fix DuplicateNodeRemover in convert pipeline (#3875)
Prior to this change calls to the `convert` command with `DuplicateNodeRemover` in `convert.ops` would result in a map read error due to `map.merge.ignore.duplicate.ids=false`. This change ensures that `map.merge.ignore.duplicate.ids=true` when `DuplicateNodeRemover` is in the `convert` pipeline.

Also added the same change to the `conflate` pipeline for consistency, even though `DuplicateNodeRemover` seemed to be working there already.