3895 - Prevent sidewalks from conflating with roads (#3902)
Made the tag similarity threshold required for road matching more strict when sidewalks are involved. The value may need additional tweaking. The more appropriate way to handle this would have been in the schema, but as noted in the code, due to the large amount of interrelations between highway types in the schema I couldn't get that to work. May be able to go back later and move the change to the schema instead.

Also discovered a related issue where `footway=crossing` features get snapped incorrectly and opened #3901 for it.