3896 - Add auto-retry for Alpha Shape generation when initial alpha too small (#3908)
After this change, `AlphaShapeGenerator` will try a second time to generate an alpha shape if the generation fails the first time when the alpha value is too small. It will do this only if any alpha shape edge face size is larger than the initial alpha value and use the largest face size for the alpha value.