3906 - Flag roads for review that cross select polygons after conflation (#3930)
Adds a post conflate step that flags for review any roads crossing polygon features according to a specified set of rules. Details on the rule format may be found in the "Crossing Roads" section of the [validation/cleaning doc](https://github.com/ngageoint/hootenanny/blob/master/docs/user/FeatureValidationAndCleaning.asciidoc). 

These changes do not attempt to prevent or correct roads crossing through buildings as doing so would be largely dependent on which of the two input datasets is favored and would require a lot of additional work and testing. Adding such a feature may be possible, though.