3978 - Handle many to one area matches (#3980)
Area Conflation was trained primarily on one to one matches. There was a recent request to diff conflate some `landuse=residential` where the secondary data consisted of a lot of smaller areas that overlapped with reference areas in a many to one fashion. These changes flag those situations as reviews, which allows them to fall out of the diff output with the default config of `differential.treat.reviews.as.matches=true`.