4073 - Update alpha shape generation to better cover sparse clusters of points (#4087)
The alpha shape generation algorithm by default can result in a forcing a user to select a 
single large output shape in order to cover datasets with sparse, isolated clusters of points. Sometimes this isn't desirable, like when creating tasking grid, as the larger shape will contribute to a tasking grid with too few cells. There wasn't any obvious way to improve the existing algorithm via manipulation of the triangulation, so the change was made to add a post shape creation step that manually covers any uncovered points at the end. 

Care still needs to be taken in the selected alpha value to get the desired output. Added #4085 which references an idea for auto-selecting an alpha value based on the input data.

Also updated `RemoveTagsVisitor` and `KeepTagsVisitor` to support wildcard string matching.