3940 - User defined POI conflation search radii (#4096)
Prior to this change changing the search radii for a POI conflation supported type could only be done when running command line conflation by editing `Poi.js`. This change moves those search radii out into a separate file, defined by the `poi.search.radii.config` configuration option, and allows them to be overridden by the JSON config option, `poi.search.radii`. The JSON config option overrides the file. This will allow for POI conflation per feature type search radii adjustments on a per dataset basis.

Added this to the documentation after I figured it out, but this is the priority of how POI conflation determines the search radius now:
1) from the `search.radius.poi` config opt (turned off by default)
2) from the element's circular error, if its larger than what is defined in the search radius defaults
3) from the JSON in the `poi.search.radii` config option
4) from the JSON in the file pointed to by the `poi.search.radii.config` config option

Also of note is the fact, prior to this change POI conflation had separate distances defined for both matching and reviewing but actually the match distance wasn't being used at all and the review distance was being used as a search radius and had nothing to do with generating reviews. So, I renamed the review distance to just a plain distance and removed the match distances.