Allow conflation of untyped features (#4152)
This is an extension of the work that started with the generic geometry conflate scripts. Before this change, all input data to be conflated had to have a type recognized by the hoot schema. This removes that requirement, based on a use case from a user, and opens the conflation up to more generic scenarios. Hoot will log a warning if no input data has a type recognized by the schema to let the user know the generic geometry conflate scripts will be handling the data and the result may not be as good as adding a recognized type to the data that would allow it to be conflated with a specific conflation alg (e.g. roads).