Linear features not handled correctly when deriving a cut only changeset (#4160)
Recently decided to kick the tires on an initial cut only scenario using `changeset-derive-replacement`. Linear features were being removed completely outside of the cut bounds with the strict bounds interpretation when they should have been being cut at the bounds instead in the initial test, and this change fixes that. There is an additional issue with polys to be handled in #4156. This also fixed the output for a couple changeset replacement tests that weren't quite right.