Prevent runaway processing during complex line matching (#4186)
Maximal subline matching performance suffers greatly for certain input data. Previously, this was only seen with rivers but now also with generic lines (in this case possibly untagged rivers). Since it often performs very well, the current solution was to use it by default and then kick over to Frechet matching if rivers were over a certain size or node count, since it isn't always as accurate but is much faster in general. 

Size/node count has proven not to be the best metric for this, and this has been reworked to kick out of maximal matching after a configurable recursive matching call limit is reached. This capability has been added to both rivers and generic lines (haven't seen any need to do it for roads yet). This has taken the runtime for a conflate job that was seen running for 48 hours with no apparent finishing time down to ~30 minutes. 

Out of the two river regression tests we have, one was slightly negatively affected (-1%) and the other was significantly positively affected (+5%).

Unrelated: code for doing adjacent C&R operations, as well as some node density tile calc bug fixes are included.