C&R orphaned nodes and way snap fixes via better element ID preservation (#4202)
The to replace and replacement maps used in Cut & Replace must be loaded separately, with the replacement map element IDs remapped, in order to support combining the maps back together for cookie cutting. This contributes to generating create/delete changeset statements where modify statements would be preferrable. These changes add a step just prior to changeset derivation that remap the replacement map's element IDs back to those of the to replace map in instances where the elements between the two are identical.