Add capability to generate uniform task grid for C&R testing (#4215)
This adds the capability to generate a uniform task grid in addition to the existing node density based task grid, which makes C&R testing easier (also added capability to read bounds file and generate a task grid). Combined this and node density from the `node-density-tiles` command and a new `task-grid` command.