Fix C&R handling of relations with mixed geometry type children (#4289)
The previous version of C&R had the Achilles heel that it could drop or duplicate features of different geometry types that belonged to the same relation (e.g. roads in a relation with road signs). This was due to the separate processing of features with different geometries. The default implementation has been changed to process all features in a single pass to correct the problem.