Scale the maximal subline match recursion based on river input data (#4315)
This is a follow on to a previous change where `MaximalSubline` had the number of recursions executed by its match method capped at a pre-determined value when used with River Conflation. This change scales that capping based on the total length of rivers in the input data and strikes a better balance between river conflated output quality and runtime performance. The capability can be toggled using the `waterway.maximal.subline.auto.optimize` configuration option from the command line or the "Automatically optimize runtime performance" setting in the River section of the UI advanced options.