Remove logic that prevents POIs with conflated status from being merged with polys (#4329)
I think the original logic was based on the false assumption that POIs would always come in with a 1/2 status. If they are merged beforehand with a poly, that must be true b/c core requires that POIs must always be merged into polys. However, if a POI is immediately merged beforehand with another POI, it could come into poi/poly merging with a conflated status.