Add alternative way to handle poi/poly matches conflicting with poi/poi matches (#4349)
The default behavior is to auto-merge poi/polygon matches when they contain features also in a poi/poi match, as the performance
for poi/poly is often better when doing so. This adds the option, `poi.polygon.match.takes.precedence.over.poi.to.poi.review`, to allow the behavior to more closely mimic that of other intra-dataset matching where overlapping matches always result in reviews for involved features.