Improve many to one area matching (#4386)
This expands upon the changes from #3978 where we marked many to one `landuse=residential` for review, so they would fall out of a diff. This improves that by loosening the exact type tag match restriction. The use case for this was for matching lakes and reservoirs (schema relationship added), but the change applies to all area features.