Diff Conflated roads don't snap back to reference roads after changeset application (#4418)
When road snapping is enabled with Differential Conflation, this change prevents any road snapped back to from being dropped out of the diff and ensures that secondary roads in the diff snapped back to ref roads stay snapped after changeset application.  One side effect of this change is that diffs with a large amount of way snapping occurring during conflation will be larger due to the extra ref features that must be added to them. 

This also enhances the detail of the way snap debugging tags used by `UnconnectedWaySnapper` and adds a test for the three step diff workflow (diff conflate to osm --> changeset derivation --> changeset application), which is in addition to the two step diff workflow (diff conflate to changeset --> changeset application).