Fix use of bounds option by conflation when relations are fully hydrated (#4464)
This change addresses issues found in the grail workflow that performs Reference Conflation within a bounds and writes the resulting differential back to the reference data store.

Prior to this change the `conflate` command did not work properly for inputs with fully hydrated relations (generally, data source from APIDB queries). Such relations would include many members outside of the bounds specified in the `bounds` option, which `conflate` would incorrectly modify. Additionally, if there were a large number of relation members outside of the bounds, the conflate time could increase exponentially due to making unnecessary matches (especially with rivers). 

Prior to this change `changeset-derive` made no use of the `bounds` option. It now does, which is necessary when relation with members outside of the bounds are specified in order to prevent modifying ref data outside of the bounds. If `conflate` is used with `bounds`, then any following `changeset-derive` part of the same workflow should also use the same `bounds`.