Prevent conflate operations operating on generically typed elements from modifying elements that are not to be conflated (#4557)
This bug was detected when a user performed POI/Polygon conflation only on a dataset, and when a changeset was generated for the conflated data several road modifications were included (`conflate-op-4550` tests for this now specifically). 

Prior to this change, conflation would check each pre/post conflate op to see which types of features it operated on (`FilteredByGeometryTypeCriteria`). If there were no matchers in the conflate configuration associated with the feature types being operated on, the op would be removed from the conflate pipeline (`SuperfluousConflateOpRemover`). However, for ops that operated on generically typed elements (e.g. all ways) there was a bug that elements not meant to be conflated could still be modified by a conflate op.

This change adds a check to select ops that operate against generically typed elements when they are run with conflation only that verifies each feature being operated on is conflatable given the current configuration. It does this via info stored in `ConflateInfoCache` and receives the cache via `OpExecutor` (formerly `NamedOp`) through the `ConflateInfoCacheConsumer` interface.

Certain conflate ops may be exempt from implementing both `FilteredByGeometryTypeCriteria` and `ConflateInfoCacheConsumer`, and it is up to implementers to decide that.