Handle untyped and site relations better when calculating stats (#4613)
The calculation of area stats was crashing when encountering either a relation with no type or a relation of `type=site`. Now, a warning is logged instead in both cases and the offending feature excluded from the stat calculation. The `type-site` fix applies to all stats encountering this data, as well as any code attempting to convert such a feature to a geometry. The untyped relation fix currently applies only to calculating area statistics. Other callers converting such a feature to a geometry will need to be modified to not throw an error and skip the feature in question (see `CalculateAreaVisitor`).