Multilinestring relation merge bug fixes from RnD branch (#4707)
The `LinearSnapMerger` encounters situations where it will leave behind multilinestring relations for features it merges. Most of the time this is undesirable, and this change puts the `MultilineStringMergeRelationCollapser` into the conflate pipeline to remove them. Eventually, we could investigate if such relations are ever needed in the output and prevent them from being created by `LinearSnapMerger` at all.

Other changes in this: 
* Added cleaning for explicit removal of empty and zero length ways
* Updated `MaximalSublineStringMatcher` to not throw an error when encountering invalid ways; just skips them instead