Run to OGR translations streaming by default (#4718)
Prior to this change all conversions with an OGR format as an output would run translation in a separate thread. Doing so is memory bound. This change makes running translations in separate thread optional and disabled by default (`convert.translate.multithreaded`). If both input *and* output format are OGR formats, however, multi-threaded translation is still optional but the operation will always be memory bound.