Backup subline matching part 1 (#4803)
River Conflation uses a workflow where and initial matcher deemed of high accuracy but sometimes computationally expensive is used first, and then we switch to a slightly less accurate less computationally expensive subline matcher after a certain amount of processing is performed. Due to recently finding some road input data that expensive to match, this approach has now been replicated to linear matching for all features types. In a follow on PR, some changes will be made to abstract some of the `SubineStringMatcher` initialization out of the `MatchCreator` classes and js conflate scripts.