Update `compare` command to work with features other than roads (#4842)
The `compare` command has always been hardcoded to only work with roads. This change opens it up to work with a feature geometry types. Also, a filtering mechanism was added to the command to allow for only scoring the similarity of a subset of features in the inputs. The graph score is only computed against unclosed linear features, as it doesn't make sense to compute for other features.

Unrelated: Also sync'd all references to "linear waterway" to be "river", as that was inconsistent throughout.