Remove namespace prefix requirement for factory creation (#4987)
This removes the requirement that the 'hoot::' namespace be specified when creating classes via Factory for all classes except `FeatureExtractors` (see #4989). Backward compatibility is left in for clients which still use the namespace. `FeatureExtractors` have a strange dependency on their class names that causes matching quality issues when the names are changed. Initial attempts to solve this problem as part of these changes have failed.

This also remove the unused class, `PoiPolygonRfClassifier`, as POI/Polygon conflation doesn't use a model.